csm/requirements.txt
torch==2.4.0
torchaudio==2.4.0
tokenizers==0.21.0
transformers==4.49.0
huggingface_hub==0.28.1
moshi==0.2.2
torchtune==0.4.0
torchao==0.9.0
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@master
faster-whisper @ git+https://github.com/SYSTRAN/faster-whisper@master
bitsandbytes
flask_cors